move all files to the /src directory of DataPicker to start preparing this to be merged with main
