--- Dash/Project.toml
@@ -1,14 +1,0 @@
-[deps]
-Dash = "1b08a953-4be3-4667-9a23-3db579824955"
-DashBootstrapComponents = "1b08a953-4be3-4667-9a23-fa6d05876a1e"
-DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
-GeophysicalModelGenerator = "3700c31b-fa53-48a6-808a-ef22d5a84742"
-JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
-JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
-PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
-Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
-UrlDownload = "856ac37a-3032-4c1c-9122-f86d88358c8b"
-
-[compat]
-GeophysicalModelGenerator = "0.5.4"
