--- Dash/assets/colormaps/managua.txt
@@ -1,256 +1,0 @@
-1.000000 0.812630 0.404239
-0.995163 0.804546 0.401552
-0.990240 0.796488 0.398875
-0.985323 0.788476 0.396216
-0.980412 0.780499 0.393560
-0.975507 0.772566 0.390930
-0.970617 0.764681 0.388300
-0.965734 0.756837 0.385681
-0.960867 0.749039 0.383098
-0.956008 0.741291 0.380517
-0.951157 0.733597 0.377947
-0.946314 0.725951 0.375393
-0.941487 0.718348 0.372860
-0.936674 0.710803 0.370341
-0.931861 0.703303 0.367841
-0.927063 0.695854 0.365358
-0.922279 0.688453 0.362886
-0.917502 0.681092 0.360425
-0.912728 0.673792 0.358001
-0.907969 0.666530 0.355581
-0.903213 0.659321 0.353163
-0.898462 0.652158 0.350776
-0.893721 0.645033 0.348388
-0.888989 0.637961 0.346011
-0.884260 0.630933 0.343667
-0.879531 0.623946 0.341336
-0.874812 0.616998 0.339018
-0.870093 0.610091 0.336701
-0.865382 0.603234 0.334418
-0.860673 0.596415 0.332135
-0.855966 0.589631 0.329871
-0.851254 0.582896 0.327604
-0.846548 0.576213 0.325360
-0.841847 0.569544 0.323151
-0.837141 0.562939 0.320941
-0.832431 0.556354 0.318738
-0.827719 0.549825 0.316560
-0.823010 0.543331 0.314381
-0.818292 0.536881 0.312217
-0.813575 0.530459 0.310103
-0.808858 0.524083 0.307958
-0.804130 0.517752 0.305871
-0.799400 0.511449 0.303754
-0.794659 0.505193 0.301670
-0.789914 0.498983 0.299621
-0.785165 0.492803 0.297568
-0.780403 0.486677 0.295528
-0.775635 0.480583 0.293510
-0.770863 0.474541 0.291527
-0.766076 0.468526 0.289539
-0.761278 0.462554 0.287559
-0.756471 0.456632 0.285607
-0.751655 0.450751 0.283687
-0.746818 0.444910 0.281780
-0.741969 0.439096 0.279884
-0.737105 0.433334 0.278008
-0.732226 0.427620 0.276161
-0.727319 0.421918 0.274303
-0.722393 0.416275 0.272473
-0.717455 0.410668 0.270694
-0.712472 0.405076 0.268909
-0.707472 0.399518 0.267116
-0.702439 0.394006 0.265378
-0.697370 0.388520 0.263665
-0.692273 0.383057 0.261940
-0.687123 0.377612 0.260249
-0.681932 0.372187 0.258573
-0.676708 0.366801 0.256922
-0.671425 0.361420 0.255294
-0.666097 0.356067 0.253672
-0.660715 0.350728 0.252079
-0.655283 0.345391 0.250488
-0.649801 0.340089 0.248954
-0.644251 0.334804 0.247423
-0.638656 0.329526 0.245904
-0.633002 0.324248 0.244421
-0.627289 0.319014 0.242975
-0.621520 0.313798 0.241574
-0.615700 0.308601 0.240165
-0.609827 0.303410 0.238814
-0.603909 0.298260 0.237523
-0.597926 0.293142 0.236234
-0.591912 0.288054 0.235002
-0.585847 0.283016 0.233773
-0.579740 0.277995 0.232633
-0.573590 0.273018 0.231548
-0.567413 0.268076 0.230475
-0.561198 0.263215 0.229479
-0.554963 0.258371 0.228573
-0.548712 0.253612 0.227693
-0.542431 0.248906 0.226886
-0.536138 0.244237 0.226157
-0.529842 0.239680 0.225479
-0.523541 0.235196 0.224869
-0.517245 0.230755 0.224361
-0.510943 0.226433 0.223950
-0.504672 0.222166 0.223627
-0.498412 0.218021 0.223394
-0.492165 0.213965 0.223254
-0.485952 0.210003 0.223213
-0.479785 0.206181 0.223277
-0.473638 0.202416 0.223448
-0.467557 0.198797 0.223730
-0.461519 0.195316 0.224127
-0.455536 0.191949 0.224655
-0.449617 0.188732 0.225338
-0.443765 0.185656 0.226155
-0.437987 0.182662 0.227082
-0.432286 0.179868 0.228171
-0.426653 0.177234 0.229378
-0.421111 0.174741 0.230767
-0.415672 0.172385 0.232317
-0.410333 0.170229 0.234010
-0.405072 0.168215 0.235897
-0.399924 0.166404 0.237935
-0.394890 0.164752 0.240137
-0.389957 0.163309 0.242542
-0.385146 0.162035 0.245119
-0.380463 0.160929 0.247920
-0.375888 0.159997 0.250871
-0.371430 0.159317 0.254031
-0.367114 0.158833 0.257377
-0.362919 0.158533 0.260924
-0.358852 0.158428 0.264665
-0.354944 0.158527 0.268624
-0.351141 0.158823 0.272756
-0.347483 0.159307 0.277108
-0.343941 0.159992 0.281641
-0.340564 0.160939 0.286361
-0.337315 0.162065 0.291308
-0.334203 0.163376 0.296415
-0.331205 0.164857 0.301701
-0.328374 0.166583 0.307186
-0.325654 0.168471 0.312840
-0.323097 0.170563 0.318670
-0.320660 0.172833 0.324649
-0.318337 0.175302 0.330794
-0.316164 0.177969 0.337101
-0.314101 0.180741 0.343535
-0.312156 0.183735 0.350114
-0.310380 0.186896 0.356818
-0.308695 0.190206 0.363627
-0.307121 0.193696 0.370558
-0.305695 0.197316 0.377601
-0.304353 0.201063 0.384726
-0.303137 0.204999 0.391950
-0.302039 0.209051 0.399237
-0.301058 0.213227 0.406613
-0.300191 0.217560 0.414037
-0.299438 0.221984 0.421509
-0.298777 0.226557 0.429039
-0.298217 0.231221 0.436600
-0.297777 0.235995 0.444189
-0.297445 0.240848 0.451793
-0.297212 0.245818 0.459408
-0.297076 0.250896 0.467029
-0.297039 0.256031 0.474647
-0.297100 0.261272 0.482250
-0.297256 0.266580 0.489830
-0.297503 0.271940 0.497403
-0.297841 0.277408 0.504931
-0.298276 0.282920 0.512422
-0.298814 0.288473 0.519871
-0.299434 0.294083 0.527282
-0.300124 0.299760 0.534628
-0.300902 0.305477 0.541914
-0.301764 0.311218 0.549148
-0.302708 0.317001 0.556297
-0.303731 0.322826 0.563394
-0.304833 0.328665 0.570396
-0.306011 0.334539 0.577332
-0.307222 0.340419 0.584183
-0.308530 0.346314 0.590950
-0.309895 0.352235 0.597634
-0.311303 0.358172 0.604232
-0.312773 0.364104 0.610733
-0.314306 0.370048 0.617145
-0.315902 0.376004 0.623471
-0.317520 0.381947 0.629693
-0.319197 0.387897 0.635834
-0.320921 0.393847 0.641880
-0.322684 0.399790 0.647834
-0.324465 0.405749 0.653703
-0.326296 0.411682 0.659482
-0.328174 0.417627 0.665174
-0.330082 0.423555 0.670790
-0.332013 0.429498 0.676316
-0.333979 0.435442 0.681762
-0.335961 0.441374 0.687145
-0.337976 0.447313 0.692459
-0.340034 0.453269 0.697691
-0.342103 0.459234 0.702875
-0.344186 0.465199 0.707993
-0.346311 0.471198 0.713058
-0.348466 0.477201 0.718079
-0.350645 0.483227 0.723047
-0.352833 0.489280 0.727979
-0.355061 0.495368 0.732876
-0.357301 0.501485 0.737730
-0.359554 0.507627 0.742563
-0.361845 0.513807 0.747363
-0.364145 0.520010 0.752135
-0.366486 0.526272 0.756887
-0.368835 0.532560 0.761617
-0.371192 0.538887 0.766331
-0.373590 0.545251 0.771030
-0.376001 0.551659 0.775706
-0.378422 0.558095 0.780371
-0.380873 0.564581 0.785026
-0.383329 0.571100 0.789663
-0.385787 0.577657 0.794291
-0.388281 0.584255 0.798904
-0.390784 0.590884 0.803505
-0.393288 0.597553 0.808099
-0.395819 0.604255 0.812679
-0.398349 0.610985 0.817253
-0.400900 0.617743 0.821819
-0.403444 0.624542 0.826372
-0.406005 0.631368 0.830919
-0.408564 0.638221 0.835455
-0.411139 0.645102 0.839990
-0.413720 0.652022 0.844511
-0.416306 0.658958 0.849030
-0.418896 0.665934 0.853537
-0.421489 0.672938 0.858048
-0.424096 0.679960 0.862559
-0.426709 0.687024 0.867055
-0.429323 0.694108 0.871560
-0.431953 0.701233 0.876058
-0.434566 0.708388 0.880557
-0.437202 0.715574 0.885061
-0.439832 0.722782 0.889564
-0.442476 0.730039 0.894073
-0.445116 0.737321 0.898583
-0.447756 0.744641 0.903104
-0.450420 0.752000 0.907629
-0.453081 0.759390 0.912156
-0.455744 0.766824 0.916696
-0.458412 0.774293 0.921240
-0.461087 0.781809 0.925798
-0.463768 0.789358 0.930364
-0.466450 0.796943 0.934936
-0.469145 0.804576 0.939524
-0.471834 0.812242 0.944119
-0.474525 0.819949 0.948721
-0.477215 0.827698 0.953338
-0.479917 0.835486 0.957959
-0.482610 0.843308 0.962585
-0.485303 0.851165 0.967223
-0.488011 0.859063 0.971865
-0.490708 0.866990 0.976509
-0.493386 0.874952 0.981161
-0.496072 0.882941 0.985815
-0.498765 0.890960 0.990469
-0.501444 0.899006 0.995121
-0.504115 0.907076 0.999779
