--- Dash/assets/colormaps/oleron.txt
@@ -1,256 +1,0 @@
-0.101051 0.150029 0.350268
-0.107211 0.155787 0.356085
-0.113288 0.161590 0.361920
-0.119268 0.167394 0.367765
-0.125252 0.173224 0.373649
-0.131224 0.179109 0.379544
-0.137168 0.185025 0.385457
-0.143104 0.190934 0.391412
-0.149035 0.196917 0.397368
-0.154983 0.202879 0.403363
-0.160954 0.208894 0.409372
-0.166940 0.214922 0.415405
-0.172908 0.221015 0.421458
-0.178936 0.227090 0.427556
-0.184980 0.233182 0.433649
-0.191019 0.239337 0.439789
-0.197125 0.245483 0.445936
-0.203225 0.251694 0.452129
-0.209359 0.257880 0.458321
-0.215526 0.264132 0.464556
-0.221710 0.270392 0.470798
-0.227941 0.276664 0.477065
-0.234173 0.282966 0.483354
-0.240419 0.289287 0.489668
-0.246722 0.295618 0.496005
-0.253046 0.301978 0.502370
-0.259395 0.308379 0.508757
-0.265751 0.314779 0.515151
-0.272115 0.321218 0.521572
-0.278547 0.327665 0.528019
-0.284964 0.334149 0.534485
-0.291439 0.340636 0.540969
-0.297900 0.347155 0.547475
-0.304404 0.353672 0.554003
-0.310938 0.360224 0.560549
-0.317471 0.366805 0.567119
-0.324030 0.373391 0.573700
-0.330615 0.380003 0.580307
-0.337229 0.386625 0.586927
-0.343842 0.393278 0.593573
-0.350503 0.399944 0.600234
-0.357169 0.406637 0.606920
-0.363837 0.413344 0.613610
-0.370538 0.420064 0.620339
-0.377270 0.426807 0.627069
-0.384009 0.433558 0.633829
-0.390772 0.440340 0.640598
-0.397548 0.447135 0.647388
-0.404341 0.453950 0.654195
-0.411161 0.460775 0.661019
-0.417992 0.467630 0.667869
-0.424850 0.474498 0.674725
-0.431726 0.481373 0.681598
-0.438599 0.488279 0.688501
-0.445502 0.495192 0.695411
-0.452438 0.502125 0.702337
-0.459371 0.509076 0.709288
-0.466322 0.516030 0.716250
-0.473287 0.523024 0.723220
-0.480286 0.530020 0.730221
-0.487295 0.537040 0.737228
-0.494317 0.544063 0.744254
-0.501356 0.551111 0.751301
-0.508414 0.558166 0.758354
-0.515471 0.565243 0.765425
-0.522562 0.572346 0.772503
-0.529669 0.579457 0.779599
-0.536791 0.586575 0.786710
-0.543917 0.593716 0.793824
-0.551067 0.600867 0.800939
-0.558223 0.608041 0.808062
-0.565400 0.615214 0.815181
-0.572603 0.622409 0.822297
-0.579808 0.629615 0.829395
-0.587015 0.636838 0.836468
-0.594243 0.644054 0.843507
-0.601473 0.651290 0.850506
-0.608719 0.658519 0.857445
-0.615947 0.665751 0.864315
-0.623183 0.672978 0.871085
-0.630403 0.680180 0.877748
-0.637601 0.687375 0.884278
-0.644770 0.694530 0.890649
-0.651914 0.701653 0.896848
-0.658997 0.708735 0.902842
-0.666031 0.715748 0.908610
-0.672996 0.722680 0.914138
-0.679866 0.729543 0.919396
-0.686656 0.736302 0.924380
-0.693340 0.742971 0.929074
-0.699907 0.749512 0.933470
-0.706366 0.755950 0.937574
-0.712688 0.762259 0.941373
-0.718898 0.768446 0.944891
-0.724986 0.774514 0.948123
-0.730942 0.780465 0.951099
-0.736792 0.786306 0.953828
-0.742535 0.792032 0.956334
-0.748176 0.797664 0.958640
-0.753726 0.803211 0.960767
-0.759197 0.808685 0.962737
-0.764606 0.814084 0.964572
-0.769952 0.819431 0.966298
-0.775252 0.824734 0.967932
-0.780515 0.830002 0.969481
-0.785750 0.835235 0.970971
-0.790956 0.840451 0.972415
-0.796148 0.845644 0.973813
-0.801330 0.850830 0.975183
-0.806509 0.856015 0.976533
-0.811679 0.861191 0.977867
-0.816850 0.866371 0.979190
-0.822025 0.871554 0.980503
-0.827199 0.876733 0.981812
-0.832379 0.881927 0.983114
-0.837567 0.887121 0.984419
-0.842756 0.892326 0.985718
-0.847948 0.897534 0.987020
-0.853150 0.902750 0.988314
-0.858355 0.907974 0.989617
-0.863572 0.913201 0.990914
-0.868787 0.918435 0.992213
-0.873999 0.923682 0.993511
-0.879219 0.928927 0.994803
-0.884444 0.934184 0.996097
-0.889665 0.939450 0.997391
-0.894892 0.944721 0.998686
-0.900113 0.949993 0.999975
-0.100240 0.299006 0.000155
-0.112059 0.301986 0.000226
-0.123174 0.304970 0.000257
-0.133900 0.307897 0.000247
-0.144183 0.310824 0.000204
-0.154116 0.313690 0.000161
-0.163828 0.316526 0.000125
-0.173231 0.319321 0.000097
-0.182452 0.322081 0.000077
-0.191517 0.324797 0.000069
-0.200403 0.327489 0.000075
-0.209187 0.330157 0.000098
-0.217842 0.332794 0.000144
-0.226396 0.335398 0.000219
-0.234859 0.337967 0.000330
-0.243210 0.340551 0.000490
-0.251560 0.343129 0.000711
-0.259831 0.345692 0.001011
-0.268057 0.348297 0.001410
-0.276308 0.350930 0.001933
-0.284508 0.353583 0.002608
-0.292741 0.356317 0.003469
-0.300991 0.359099 0.004557
-0.309305 0.361990 0.005916
-0.317630 0.364963 0.007597
-0.326018 0.368063 0.009658
-0.334508 0.371283 0.012367
-0.343052 0.374663 0.015348
-0.351666 0.378181 0.018930
-0.360377 0.381864 0.023166
-0.369187 0.385710 0.028136
-0.378064 0.389735 0.033895
-0.387015 0.393928 0.040800
-0.396050 0.398292 0.047888
-0.405124 0.402804 0.055327
-0.414248 0.407471 0.063020
-0.423394 0.412283 0.070789
-0.432574 0.417212 0.078726
-0.441738 0.422234 0.086949
-0.450884 0.427379 0.095228
-0.460001 0.432585 0.103618
-0.469095 0.437854 0.112113
-0.478121 0.443165 0.120649
-0.487104 0.448515 0.129315
-0.496012 0.453885 0.138052
-0.504872 0.459278 0.146755
-0.513661 0.464677 0.155552
-0.522375 0.470077 0.164360
-0.531040 0.475480 0.173148
-0.539637 0.480870 0.181959
-0.548199 0.486280 0.190798
-0.556703 0.491704 0.199617
-0.565164 0.497122 0.208471
-0.573613 0.502563 0.217312
-0.582031 0.508043 0.226158
-0.590441 0.513543 0.234999
-0.598853 0.519090 0.243806
-0.607272 0.524679 0.252670
-0.615690 0.530332 0.261518
-0.624143 0.536045 0.270391
-0.632612 0.541829 0.279258
-0.641111 0.547695 0.288120
-0.649657 0.553632 0.297027
-0.658233 0.559662 0.305938
-0.666846 0.565761 0.314833
-0.675508 0.571955 0.323770
-0.684205 0.578221 0.332732
-0.692958 0.584577 0.341682
-0.701731 0.591002 0.350664
-0.710557 0.597505 0.359644
-0.719414 0.604078 0.368655
-0.728299 0.610707 0.377669
-0.737219 0.617396 0.386700
-0.746161 0.624147 0.395771
-0.755130 0.630945 0.404836
-0.764105 0.637785 0.413942
-0.773094 0.644669 0.423061
-0.782091 0.651605 0.432233
-0.791075 0.658568 0.441410
-0.800051 0.665577 0.450634
-0.808995 0.672625 0.459894
-0.817893 0.679697 0.469215
-0.826745 0.686819 0.478566
-0.835514 0.693969 0.487989
-0.844186 0.701164 0.497461
-0.852739 0.708393 0.506988
-0.861136 0.715654 0.516582
-0.869359 0.722932 0.526244
-0.877367 0.730249 0.535948
-0.885134 0.737573 0.545718
-0.892625 0.744911 0.555514
-0.899805 0.752247 0.565352
-0.906647 0.759568 0.575215
-0.913125 0.766871 0.585083
-0.919214 0.774134 0.594931
-0.924907 0.781355 0.604762
-0.930195 0.788513 0.614543
-0.935066 0.795601 0.624286
-0.939537 0.802617 0.633958
-0.943612 0.809557 0.643549
-0.947313 0.816413 0.653068
-0.950661 0.823182 0.662520
-0.953686 0.829879 0.671879
-0.956415 0.836497 0.681164
-0.958882 0.843049 0.690387
-0.961118 0.849544 0.699548
-0.963155 0.855984 0.708665
-0.965019 0.862387 0.717732
-0.966746 0.868754 0.726764
-0.968355 0.875091 0.735776
-0.969864 0.881418 0.744780
-0.971299 0.887731 0.753777
-0.972677 0.894046 0.762774
-0.974000 0.900360 0.771786
-0.975289 0.906683 0.780806
-0.976553 0.913014 0.789846
-0.977795 0.919355 0.798909
-0.979018 0.925717 0.807992
-0.980230 0.932088 0.817101
-0.981427 0.938483 0.826236
-0.982613 0.944895 0.835399
-0.983785 0.951319 0.844588
-0.984948 0.957763 0.853798
-0.986093 0.964218 0.863044
-0.987224 0.970694 0.872308
-0.988330 0.977184 0.881595
-0.989424 0.983687 0.890903
-0.990487 0.990203 0.900235
