--- Dash/assets/colormaps/vik.txt
@@ -1,256 +1,0 @@
-0.001328 0.069836 0.379529
-0.002366 0.076475 0.383518
-0.003304 0.083083 0.387487
-0.004146 0.089590 0.391477
-0.004897 0.095948 0.395453
-0.005563 0.102274 0.399409
-0.006151 0.108500 0.403388
-0.006668 0.114686 0.407339
-0.007119 0.120845 0.411288
-0.007512 0.126958 0.415230
-0.007850 0.133068 0.419166
-0.008141 0.139092 0.423079
-0.008391 0.145171 0.427006
-0.008606 0.151144 0.430910
-0.008790 0.157140 0.434809
-0.008947 0.163152 0.438691
-0.009080 0.169142 0.442587
-0.009193 0.175103 0.446459
-0.009290 0.181052 0.450337
-0.009372 0.187051 0.454212
-0.009443 0.193028 0.458077
-0.009506 0.198999 0.461951
-0.009564 0.205011 0.465816
-0.009619 0.211021 0.469707
-0.009675 0.217047 0.473571
-0.009735 0.223084 0.477461
-0.009802 0.229123 0.481352
-0.009881 0.235206 0.485250
-0.009977 0.241277 0.489161
-0.010098 0.247386 0.493080
-0.010254 0.253516 0.497020
-0.010463 0.259675 0.500974
-0.010755 0.265853 0.504938
-0.011176 0.272037 0.508925
-0.011716 0.278296 0.512923
-0.012286 0.284554 0.516953
-0.012934 0.290865 0.520998
-0.013790 0.297214 0.525074
-0.014838 0.303577 0.529184
-0.016131 0.310015 0.533308
-0.017711 0.316474 0.537485
-0.019630 0.322986 0.541677
-0.021948 0.329550 0.545931
-0.024730 0.336144 0.550210
-0.028047 0.342826 0.554538
-0.031980 0.349543 0.558906
-0.036812 0.356332 0.563341
-0.042229 0.363171 0.567811
-0.048008 0.370086 0.572345
-0.054292 0.377080 0.576933
-0.060963 0.384129 0.581571
-0.068081 0.391265 0.586280
-0.075457 0.398460 0.591042
-0.083246 0.405740 0.595868
-0.091425 0.413088 0.600754
-0.099832 0.420499 0.605697
-0.108595 0.428000 0.610711
-0.117694 0.435566 0.615770
-0.127042 0.443194 0.620895
-0.136702 0.450888 0.626062
-0.146607 0.458643 0.631289
-0.156787 0.466457 0.636560
-0.167187 0.474324 0.641866
-0.177807 0.482238 0.647218
-0.188606 0.490191 0.652599
-0.199580 0.498193 0.658021
-0.210783 0.506201 0.663465
-0.222120 0.514263 0.668924
-0.233602 0.522322 0.674403
-0.245231 0.530414 0.679894
-0.256999 0.538517 0.685405
-0.268867 0.546617 0.690908
-0.280797 0.554717 0.696428
-0.292852 0.562822 0.701935
-0.304985 0.570907 0.707448
-0.317174 0.578997 0.712950
-0.329438 0.587064 0.718447
-0.341729 0.595123 0.723934
-0.354067 0.603164 0.729412
-0.366459 0.611186 0.734877
-0.378862 0.619189 0.740325
-0.391305 0.627159 0.745757
-0.403760 0.635114 0.751183
-0.416227 0.643046 0.756582
-0.428711 0.650956 0.761968
-0.441199 0.658836 0.767341
-0.453697 0.666696 0.772699
-0.466195 0.674537 0.778044
-0.478697 0.682349 0.783369
-0.491208 0.690143 0.788682
-0.503691 0.697910 0.793980
-0.516178 0.705661 0.799260
-0.528677 0.713387 0.804525
-0.541149 0.721090 0.809775
-0.553624 0.728778 0.815010
-0.566096 0.736441 0.820229
-0.578557 0.744089 0.825435
-0.591014 0.751718 0.830626
-0.603468 0.759314 0.835793
-0.615908 0.766896 0.840941
-0.628351 0.774452 0.846058
-0.640779 0.781988 0.851147
-0.653203 0.789485 0.856206
-0.665631 0.796945 0.861214
-0.678051 0.804371 0.866172
-0.690457 0.811742 0.871059
-0.702868 0.819048 0.875866
-0.715265 0.826290 0.880567
-0.727646 0.833439 0.885146
-0.740019 0.840479 0.889570
-0.752354 0.847380 0.893807
-0.764662 0.854125 0.897821
-0.776918 0.860678 0.901565
-0.789096 0.866991 0.904992
-0.801170 0.873031 0.908043
-0.813110 0.878738 0.910653
-0.824870 0.884062 0.912761
-0.836396 0.888934 0.914302
-0.847617 0.893289 0.915195
-0.858470 0.897074 0.915385
-0.868874 0.900206 0.914812
-0.878729 0.902636 0.913418
-0.887965 0.904303 0.911164
-0.896497 0.905178 0.908034
-0.904242 0.905221 0.904013
-0.911151 0.904422 0.899132
-0.917175 0.902800 0.893409
-0.922285 0.900367 0.886911
-0.926482 0.897173 0.879687
-0.929789 0.893256 0.871826
-0.932236 0.888698 0.863396
-0.933880 0.883552 0.854476
-0.934782 0.877893 0.845152
-0.935013 0.871795 0.835493
-0.934644 0.865313 0.825561
-0.933752 0.858522 0.815421
-0.932408 0.851469 0.805112
-0.930682 0.844208 0.794685
-0.928622 0.836778 0.784169
-0.926298 0.829215 0.773579
-0.923752 0.821545 0.762958
-0.921017 0.813795 0.752313
-0.918147 0.805997 0.741659
-0.915156 0.798157 0.731008
-0.912080 0.790294 0.720370
-0.908933 0.782421 0.709752
-0.905741 0.774540 0.699150
-0.902506 0.766670 0.688588
-0.899249 0.758812 0.678051
-0.895973 0.750973 0.667550
-0.892690 0.743148 0.657086
-0.889402 0.735345 0.646657
-0.886118 0.727569 0.636274
-0.882831 0.719826 0.625923
-0.879556 0.712106 0.615618
-0.876289 0.704419 0.605357
-0.873033 0.696764 0.595141
-0.869784 0.689144 0.584972
-0.866551 0.681541 0.574832
-0.863333 0.673985 0.564746
-0.860121 0.666453 0.554708
-0.856920 0.658957 0.544709
-0.853732 0.651500 0.534753
-0.850562 0.644061 0.524842
-0.847402 0.636670 0.514974
-0.844258 0.629296 0.505146
-0.841125 0.621957 0.495369
-0.838005 0.614653 0.485627
-0.834895 0.607392 0.475941
-0.831802 0.600144 0.466284
-0.828715 0.592938 0.456675
-0.825639 0.585758 0.447109
-0.822582 0.578600 0.437595
-0.819528 0.571478 0.428106
-0.816496 0.564388 0.418657
-0.813463 0.557328 0.409260
-0.810446 0.550285 0.399892
-0.807443 0.543274 0.390575
-0.804446 0.536288 0.381299
-0.801454 0.529329 0.372040
-0.798475 0.522380 0.362835
-0.795500 0.515460 0.353660
-0.792535 0.508575 0.344523
-0.789573 0.501692 0.335435
-0.786617 0.494827 0.326343
-0.783657 0.487977 0.317312
-0.780695 0.481123 0.308300
-0.777737 0.474295 0.299327
-0.774763 0.467464 0.290352
-0.771788 0.460620 0.281424
-0.768787 0.453783 0.272508
-0.765776 0.446929 0.263640
-0.762724 0.440055 0.254764
-0.759638 0.433147 0.245872
-0.756510 0.426200 0.237047
-0.753316 0.419216 0.228190
-0.750051 0.412163 0.219330
-0.746698 0.405028 0.210470
-0.743239 0.397819 0.201593
-0.739651 0.390493 0.192739
-0.735899 0.383060 0.183852
-0.731988 0.375473 0.174977
-0.727865 0.367743 0.166045
-0.723516 0.359852 0.157131
-0.718915 0.351766 0.148211
-0.714028 0.343503 0.139282
-0.708841 0.335048 0.130458
-0.703318 0.326354 0.121545
-0.697448 0.317502 0.112841
-0.691227 0.308462 0.104132
-0.684653 0.299264 0.095633
-0.677734 0.289916 0.087350
-0.670476 0.280477 0.079197
-0.662904 0.271015 0.071510
-0.655048 0.261520 0.064079
-0.646969 0.252081 0.057104
-0.638686 0.242711 0.050618
-0.630261 0.233488 0.044750
-0.621722 0.224449 0.039414
-0.613135 0.215657 0.034829
-0.604539 0.207086 0.031072
-0.595947 0.198741 0.028212
-0.587403 0.190700 0.026019
-0.578937 0.182918 0.024396
-0.570545 0.175423 0.023257
-0.562268 0.168171 0.022523
-0.554076 0.161202 0.022110
-0.546007 0.154400 0.021861
-0.538043 0.147854 0.021737
-0.530182 0.141491 0.021722
-0.522424 0.135276 0.021800
-0.514776 0.129209 0.021957
-0.507213 0.123272 0.022179
-0.499733 0.117487 0.022455
-0.492348 0.111818 0.022775
-0.485034 0.106209 0.023130
-0.477801 0.100607 0.023513
-0.470639 0.095156 0.023916
-0.463530 0.089668 0.024336
-0.456494 0.084258 0.024766
-0.449521 0.078741 0.025203
-0.442603 0.073404 0.025644
-0.435737 0.067904 0.026084
-0.428918 0.062415 0.026522
-0.422146 0.056832 0.026954
-0.415437 0.051116 0.027378
-0.408768 0.045352 0.027790
-0.402132 0.039448 0.028189
-0.395562 0.033385 0.028570
-0.389015 0.027844 0.028932
-0.382496 0.022586 0.029271
-0.376028 0.017608 0.029583
-0.369578 0.012890 0.029866
-0.363161 0.008243 0.030115
-0.356785 0.004035 0.030327
-0.350423 0.000061 0.030499
