+++ Project.toml
@@ -1,16 +1,31 @@
 name = "DataPicker"
-uuid = "d0d4e46c-39cb-4708-b3c6-d9d4047ce5e1"
-authors = ["Marcel Thielman <Marcel.Thielmann@uni-bayreuth.de>", "Boris Kaus <kaus@uni-mainz.de>"]
+uuid = "ad26291b-5ad9-4b61-8cdb-f4ecfd1c1fc8"
+authors = ["Boris Kaus <kaus@uni-mainz.de>", "Marcel Thielmann <Marcel.Thielmann@uni-bayreuth.de>"]
 version = "0.1.0"
 
 [deps]
-FastMarching = "7c16e180-9f04-11e8-24a6-e7c7f74617b0"
-FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
-GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
-GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
-LightXML = "9c8b4983-aa76-5018-a973-4c85ecc9e179"
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-NearestNeighbors = "b8a86587-4115-5ab1-83bc-aa920d37bbce"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Base64 = "2a0f44e3-6c83-55bd-87e4-b1978d98bd5f"
+Dash = "1b08a953-4be3-4667-9a23-3db579824955"
+DashBootstrapComponents = "1b08a953-4be3-4667-9a23-fa6d05876a1e"
+DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
+GeophysicalModelGenerator = "3700c31b-fa53-48a6-808a-ef22d5a84742"
+JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
+JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
+PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
+Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
+UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
+UrlDownload = "856ac37a-3032-4c1c-9122-f86d88358c8b"
+
+[compat]
+Dash = "1.3.0"
+DashBootstrapComponents = "1.4.1"
+DelimitedFiles = "1"
+JLD2 = "0.4"
+JSON3 = "1"
+PlotlyJS = "0.18"
+Statistics = "1.9"
+UrlDownload = "1.0.1"
+GeophysicalModelGenerator = "0.5.4"
+julia = "1.9"
