+++ src/GMG_Tomo/assets/colormaps/broc.txt
@@ -1,0 +1,256 @@
+0.170813 0.100272 0.299848
+0.170748 0.106342 0.305576
+0.170609 0.112289 0.311297
+0.170403 0.118210 0.317049
+0.170134 0.124041 0.322824
+0.169811 0.129949 0.328601
+0.169435 0.135723 0.334405
+0.168997 0.141581 0.340213
+0.168498 0.147385 0.346028
+0.167974 0.153166 0.351874
+0.167445 0.159000 0.357753
+0.166877 0.164801 0.363613
+0.166257 0.170632 0.369507
+0.165615 0.176451 0.375412
+0.164988 0.182307 0.381350
+0.164375 0.188220 0.387280
+0.163729 0.194112 0.393244
+0.163063 0.199986 0.399218
+0.162410 0.205964 0.405225
+0.161776 0.211935 0.411241
+0.161137 0.217920 0.417286
+0.160495 0.223934 0.423329
+0.159913 0.229986 0.429411
+0.159401 0.236112 0.435511
+0.158950 0.242231 0.441620
+0.158561 0.248420 0.447736
+0.158251 0.254627 0.453885
+0.158049 0.260865 0.460036
+0.157989 0.267147 0.466202
+0.158100 0.273510 0.472379
+0.158406 0.279890 0.478549
+0.158920 0.286307 0.484715
+0.159665 0.292797 0.490895
+0.160748 0.299339 0.497035
+0.162160 0.305921 0.503159
+0.163910 0.312516 0.509272
+0.165993 0.319192 0.515336
+0.168538 0.325884 0.521367
+0.171487 0.332648 0.527351
+0.174901 0.339419 0.533271
+0.178739 0.346199 0.539135
+0.182999 0.353034 0.544930
+0.187764 0.359870 0.550650
+0.192881 0.366728 0.556286
+0.198422 0.373584 0.561860
+0.204385 0.380450 0.567340
+0.210653 0.387302 0.572743
+0.217273 0.394164 0.578054
+0.224158 0.401030 0.583302
+0.231360 0.407862 0.588471
+0.238748 0.414696 0.593581
+0.246367 0.421515 0.598619
+0.254181 0.428339 0.603610
+0.262132 0.435152 0.608546
+0.270240 0.441951 0.613423
+0.278444 0.448742 0.618281
+0.286742 0.455526 0.623102
+0.295150 0.462307 0.627893
+0.303628 0.469101 0.632663
+0.312163 0.475882 0.637417
+0.320788 0.482664 0.642159
+0.329439 0.489447 0.646891
+0.338110 0.496239 0.651618
+0.346851 0.503037 0.656342
+0.355617 0.509847 0.661055
+0.364396 0.516658 0.665779
+0.373217 0.523479 0.670506
+0.382061 0.530301 0.675225
+0.390941 0.537140 0.679949
+0.399821 0.543973 0.684683
+0.408741 0.550822 0.689427
+0.417677 0.557680 0.694160
+0.426626 0.564541 0.698909
+0.435605 0.571413 0.703666
+0.444595 0.578292 0.708426
+0.453590 0.585191 0.713186
+0.462609 0.592090 0.717959
+0.471661 0.598991 0.722729
+0.480705 0.605899 0.727513
+0.489775 0.612823 0.732306
+0.498868 0.619763 0.737095
+0.507973 0.626697 0.741893
+0.517088 0.633649 0.746699
+0.526218 0.640603 0.751514
+0.535355 0.647566 0.756326
+0.544519 0.654537 0.761149
+0.553689 0.661524 0.765979
+0.562886 0.668514 0.770814
+0.572084 0.675509 0.775651
+0.581296 0.682514 0.780498
+0.590525 0.689536 0.785353
+0.599774 0.696557 0.790209
+0.609036 0.703589 0.795074
+0.618302 0.710631 0.799949
+0.627585 0.717684 0.804822
+0.636886 0.724741 0.809705
+0.646192 0.731807 0.814590
+0.655519 0.738885 0.819482
+0.664852 0.745965 0.824383
+0.674206 0.753055 0.829293
+0.683568 0.760160 0.834202
+0.692951 0.767267 0.839111
+0.702332 0.774383 0.844025
+0.711736 0.781512 0.848944
+0.721141 0.788642 0.853854
+0.730566 0.795773 0.858767
+0.740000 0.802911 0.863676
+0.749428 0.810051 0.868568
+0.758875 0.817191 0.873433
+0.768316 0.824324 0.878277
+0.777764 0.831455 0.883079
+0.787200 0.838559 0.887829
+0.796616 0.845645 0.892512
+0.806017 0.852703 0.897100
+0.815381 0.859711 0.901565
+0.824692 0.866656 0.905885
+0.833942 0.873524 0.910014
+0.843089 0.880287 0.913910
+0.852117 0.886920 0.917522
+0.860974 0.893377 0.920781
+0.869624 0.899632 0.923647
+0.878009 0.905622 0.926025
+0.886061 0.911294 0.927850
+0.893702 0.916594 0.929054
+0.900863 0.921454 0.929562
+0.907460 0.925812 0.929310
+0.913404 0.929602 0.928248
+0.918623 0.932772 0.926345
+0.923067 0.935283 0.923586
+0.926673 0.937103 0.919971
+0.929429 0.938215 0.915537
+0.931330 0.938632 0.910342
+0.932392 0.938377 0.904431
+0.932660 0.937487 0.897896
+0.932188 0.936014 0.890812
+0.931046 0.934007 0.883261
+0.929296 0.931540 0.875321
+0.927022 0.928665 0.867058
+0.924297 0.925446 0.858541
+0.921186 0.921935 0.849823
+0.917759 0.918178 0.840943
+0.914062 0.914224 0.831943
+0.910150 0.910107 0.822845
+0.906063 0.905860 0.813677
+0.901837 0.901508 0.804466
+0.897507 0.897082 0.795212
+0.893086 0.892588 0.785938
+0.888605 0.888045 0.776645
+0.884072 0.883469 0.767334
+0.879495 0.878860 0.758028
+0.874893 0.874230 0.748712
+0.870261 0.869582 0.739405
+0.865614 0.864915 0.730086
+0.860943 0.860235 0.720766
+0.856252 0.855534 0.711456
+0.851534 0.850804 0.702131
+0.846790 0.846055 0.692818
+0.842017 0.841276 0.683482
+0.837197 0.836452 0.674142
+0.832324 0.831583 0.664785
+0.827389 0.826647 0.655416
+0.822383 0.821640 0.646020
+0.817280 0.816546 0.636606
+0.812078 0.811349 0.627148
+0.806758 0.806033 0.617663
+0.801293 0.800585 0.608152
+0.795679 0.794985 0.598591
+0.789900 0.789225 0.589002
+0.783941 0.783279 0.579401
+0.777783 0.777151 0.569761
+0.771426 0.770818 0.560132
+0.764862 0.764282 0.550498
+0.758093 0.757545 0.540889
+0.751125 0.750609 0.531329
+0.743954 0.743479 0.521826
+0.736605 0.736161 0.512424
+0.729094 0.728687 0.503123
+0.721431 0.721062 0.493968
+0.713649 0.713314 0.484948
+0.705759 0.705457 0.476109
+0.697780 0.697511 0.467440
+0.689747 0.689508 0.458953
+0.681656 0.681442 0.450653
+0.673548 0.673359 0.442544
+0.665420 0.665250 0.434605
+0.657294 0.657143 0.426845
+0.649168 0.649033 0.419245
+0.641048 0.640928 0.411788
+0.632958 0.632848 0.404468
+0.624883 0.624783 0.397274
+0.616836 0.616742 0.390185
+0.608813 0.608725 0.383206
+0.600802 0.600719 0.376299
+0.592833 0.592753 0.369453
+0.584878 0.584799 0.362677
+0.576950 0.576874 0.355964
+0.569037 0.568960 0.349284
+0.561156 0.561078 0.342655
+0.553294 0.553215 0.336042
+0.545463 0.545383 0.329488
+0.537650 0.537568 0.322942
+0.529852 0.529769 0.316424
+0.522068 0.521982 0.309933
+0.514328 0.514241 0.303442
+0.506587 0.506496 0.297009
+0.498882 0.498790 0.290561
+0.491202 0.491106 0.284151
+0.483518 0.483419 0.277769
+0.475880 0.475778 0.271386
+0.468247 0.468142 0.265028
+0.460641 0.460531 0.258705
+0.453069 0.452958 0.252389
+0.445503 0.445388 0.246078
+0.437978 0.437859 0.239808
+0.430459 0.430335 0.233535
+0.422959 0.422831 0.227320
+0.415497 0.415365 0.221113
+0.408053 0.407917 0.214890
+0.400645 0.400503 0.208725
+0.393233 0.393089 0.202566
+0.385855 0.385708 0.196466
+0.378514 0.378364 0.190337
+0.371184 0.371031 0.184261
+0.363889 0.363734 0.178237
+0.356626 0.356471 0.172181
+0.349373 0.349219 0.166190
+0.342153 0.342000 0.160207
+0.334967 0.334819 0.154299
+0.327777 0.327637 0.148403
+0.320649 0.320521 0.142577
+0.313525 0.313410 0.136759
+0.306440 0.306348 0.131029
+0.299374 0.299307 0.125311
+0.292330 0.292296 0.119650
+0.285315 0.285321 0.114061
+0.278356 0.278412 0.108536
+0.271400 0.271513 0.103142
+0.264491 0.264674 0.097785
+0.257606 0.257872 0.092474
+0.250762 0.251143 0.087347
+0.243936 0.244416 0.082262
+0.237196 0.237795 0.077192
+0.230424 0.231199 0.072349
+0.223734 0.224629 0.067628
+0.217087 0.218178 0.063025
+0.210464 0.211770 0.058419
+0.203943 0.205390 0.053781
+0.197487 0.199073 0.048969
+0.191145 0.192843 0.043546
+0.184939 0.186651 0.037943
+0.178822 0.180505 0.031907
+0.172805 0.174445 0.026044
+0.166947 0.168396 0.020145
+0.161186 0.162428 0.014202
+0.155539 0.156487 0.008001
+0.150046 0.150558 0.001930
